add auth authentication for edit ticket

+++ src/repositories/ticketEditorRepository.ts
@@ -1,7 +1,7 @@
 import { DBError } from "../controllers/DBError";
 import type { Answer } from "../models/Answer";
 import { HTTP_STATUSES } from "../utils";
-import { ticketCollection } from "./db";
+import { ticketCollection, userCollection } from "./db";
 
 type CreateQuestion = {
 	imgInfo: {
@@ -14,6 +14,7 @@ type CreateQuestion = {
 	question: string;
 	help: string;
 	answers: Answer[];
+	userId: string;
 };
 
 const findTicket = async (ticketId: string) => {
@@ -26,13 +27,42 @@ const findTicket = async (ticketId: string) => {
 	return ticket;
 };
 
+const isUserExist = async (userId: string) => {
+	const filter = { userId };
+	const user = await userCollection.findOne(filter);
+	if (!user) {
+		throw new DBError("Пользователь не найден", HTTP_STATUSES.NOT_FOUND_404);
+	}
+	return user;
+};
+
+const checkAccessByRole = (role: string) => {
+	if (role === "user") {
+		throw new DBError(
+			"У вас нет прав доступа, для смены роли",
+			HTTP_STATUSES.BAD_REQUEST_400,
+		);
+	}
+};
+
 export const ticketEditorRepository = {
-	async createTicket(ticketId: string, createdAt: number) {
+	async createTicket(data: {
+		ticketId: string;
+		createdAt: number;
+		userId: string;
+	}) {
+		const { createdAt, ticketId, userId } = data;
+		const user = await isUserExist(userId);
+		checkAccessByRole(user.role);
 		await ticketCollection.insertOne({ createdAt, ticketId, questions: [] });
 	},
 
 	async addQuestion(data: CreateQuestion) {
-		const { imgInfo, questionId, ticketId, question, help, answers } = data;
+		const { imgInfo, questionId, ticketId, question, help, answers, userId } =
+			data;
+		const user = await isUserExist(userId);
+		checkAccessByRole(user.role);
+
 		const ticket = await findTicket(ticketId);
 
 		if (ticket.questions.length >= 20) {
@@ -49,13 +79,20 @@ export const ticketEditorRepository = {
 			},
 		);
 	},
-	async getQuestionsInTicket(ticketId: string) {
+	async getQuestionsInTicket(ticketId: string, userId: string) {
+		const user = await isUserExist(userId);
+		checkAccessByRole(user.role);
+
 		const ticket = await findTicket(ticketId);
 		return ticket.questions;
 	},
 
 	async editQuestion(data: CreateQuestion) {
-		const { imgInfo, questionId, ticketId, question, help, answers } = data;
+		const { imgInfo, questionId, ticketId, question, help, answers, userId } =
+			data;
+		const user = await isUserExist(userId);
+		checkAccessByRole(user.role);
+
 		const query = {
 			ticketId: ticketId,
 			"questions.questionId": questionId,
@@ -81,8 +118,17 @@ export const ticketEditorRepository = {
 		);
 	},
 
-	async findQuestion(ticketId: string, questionId: string) {
+	async findQuestion(data: {
+		ticketId: string;
+		questionId: string;
+		userId: string;
+	}) {
 		// Поиск документа с использованием $elemMatch
+		const { ticketId, questionId, userId } = data;
+
+		const user = await isUserExist(userId);
+		checkAccessByRole(user.role);
+
 		const query = {
 			ticketId,
 			questions: {
@@ -97,20 +143,33 @@ export const ticketEditorRepository = {
 			"questions.$": 1,
 		};
 
-		const ticket = await ticketCollection.findOne(query, { projection });
+		const ticket = await ticketCollection.findOne(query, {
+			projection,
+		});
 		if (ticket?.questions && ticket.questions.length > 0) {
 			return ticket.questions[0];
 		}
 		throw new DBError("Билет не найден", HTTP_STATUSES.NOT_FOUND_404);
 	},
 
-	async deleteTicket(ticketId: string) {
+	async deleteTicket(ticketId: string, userId: string) {
+		const user = await isUserExist(userId);
+		checkAccessByRole(user.role);
+
 		await findTicket(ticketId);
 
 		await ticketCollection.deleteOne({ ticketId });
 	},
 
-	async deleteQuestion(ticketId: string, questionId: string) {
+	async deleteQuestion(data: {
+		ticketId: string;
+		questionId: string;
+		userId: string;
+	}) {
+		const { ticketId, questionId, userId } = data;
+		const user = await isUserExist(userId);
+		checkAccessByRole(user.role);
+
 		await findTicket(ticketId);
 
 		const query = {
